+++ src/en/rizzcomic/src/eu/kanade/tachiyomi/extension/en/rizzcomic/RizzComic.kt
@@ -12,6 +12,8 @@ import eu.kanade.tachiyomi.source.model.Page
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
 import eu.kanade.tachiyomi.util.asJsoup
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import okhttp3.FormBody
 import okhttp3.Request
@@ -120,6 +122,31 @@ class RizzComic : MangaThemesia(
         }
 
         return FilterList(filters)
+    }
+
+    @Serializable
+    class Comic(
+        val id: Int,
+        val title: String,
+        @SerialName("image_url") val cover: String? = null,
+        @SerialName("long_description") val synopsis: String? = null,
+        val status: String? = null,
+        val type: String? = null,
+        val artist: String? = null,
+        val author: String? = null,
+        val serialization: String? = null,
+        @SerialName("genre_id") val genres: String? = null,
+    ) {
+        val slug get() = title.trim().lowercase()
+            .replace(slugRegex, "-")
+            .replace("-s-", "s-")
+            .replace("-ll-", "ll-")
+
+        val genreIds get() = genres?.split(",")?.map(String::trim)
+
+        companion object {
+            private val slugRegex = Regex("""[^a-z0-9]+""")
+        }
     }
 
     override fun searchMangaParse(response: Response): MangasPage {
